warn when the directory does not exist

+++ lib/irb/history.rb
@@ -53,14 +53,16 @@ module IRB
           raise
         end
 
-        if !File.exist?(history_file)
-          warn "Warning: The path for the HISTORY_FILE does not exist."
-        end
-
         if File.exist?(history_file) &&
            File.mtime(history_file) != @loaded_history_mtime
           history = history[@loaded_history_lines..-1] if @loaded_history_lines
           append_history = true
+        end
+
+        pathname = Pathname.new(history_file)
+        unless Dir.exist?(pathname.dirname)
+          warn "Warning: The path to save IRB history does not exist. Please, set the env var IRB.conf[:HISTORY_FILE] with a valid path."
+          return
         end
 
         File.open(history_file, (append_history ? 'a' : 'w'), 0o600, encoding: IRB.conf[:LC_MESSAGES]&.encoding) do |f|
